fix joined table name collision on properties with same names
problem:
if class A references B as two properties, one of which is one-to-one
and the other is many-to-many, then they may have same "columnName"
in proto.

it results in failure to build correct joins when both properties are
used in criteria, due to a collision in aliases for joined tables.

fix:
change how "columnName" is generated by default for many-to-many
properties
add check for name collision for columns in table